クラスの抽出:TextureBrowserWindow
